feat: overhaul to new architecture
This mega-commit introduces the new architecture @mr0grog & I discussed,
definining initial interfaces for each component & adapting the configuration
to fit. The vast majority of this work is in the lib package, and the cmd
package has gone backwards a bunch. We'll get it back soon enough by new means.

I've intentionally left lots of bits unfinished, but focused on making golint
happy. godoc should produce something modestly reasonable for critique.

There's a single test & example config file available in the lib package that
puts everything together, it works the `NewWalk` function, which is a nice
high-level abstraction of "I give you config, then call start & it should work".

For the record, it's bad that this is one commit.