Ensure GOPATH and reduce repeated commands
I noticed we added an `install` target that repeats code from the `install-deps` target, so this uses a dependency to clean that up. While doing so, I also realized that a) we had a `require-gopath` target we didn't use and that b) it didn't work properly. It turns out you can't use control directives like this inside a target (among the many reasons I find make confusing as all get-out), but that's OK since we should probably always be ensuring `GOPATH` is set.